requirements.txt
Flask==3.0.3
fastapi==0.115.6
click>=8.1.3
numpy>=1.23.0
opencv-python>=4.6.0
pillow>=7.1.2
requests>=2.23.0
scipy>=1.4.1
torch>=1.8.0
torchvision>=0.9.0
tqdm>=4.64.0
psutil 
PyYAML
py-cpuinfo 
pandas>=1.1.4
seaborn>=0.11.0 
ultralytics-thop>=2.0.0 
gunicorn==20.0.4
gradio_client==1.5.2
gevent>1.4.0
google-cloud-storage>=1.43.0
google-generativeai==0.8.3
firebase-admin==6.6.0
python-dotenv==0.21.0
uvicorn==0.32.1
ttach==0.0.3
python-multipart